refactor(wave0d.1 #252): move sessionTitles/ pure-logic to daemon
Move electron/sessionTitles/{deciders,index}.ts and their tests into
packages/daemon/src/sessionTitles/. Both are pure (only @anthropic-ai/
claude-agent-sdk + node builtins; no electron API). Tests renamed
.test.ts → .spec.ts. Relative imports get .js extensions.

Re-export shim at electron/sessionTitles/index.ts keeps
electron/main.ts and electron/notify/bootstrap/installPipeline.ts
compiling — Wave 0d.2 (#250) deletes the shim when sessionWatcher/
migrates and these consumers move with it.

The CommonJS dynamic-import workaround in index.ts (`new Function(
'spec', 'return import(spec)')`) is no longer technically needed under
the daemon's ESM context but is left in place to keep this PR a true
file move with zero semantic change.

[REFACTOR-ONLY] no behavior change.
